Docker setup and configurations

+++ src/main/java/com/tracker/production_line_equipment_tracker/EquipmentTrackerRepository.java
@@ -2,5 +2,5 @@ package com.tracker.production_line_equipment_tracker;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipmentTrackerRepository extends JpaRepository<Equipment, Integer> {
+public interface EquipmentTrackerRepository extends JpaRepository<Equipment, Long> {
 }
